Adding non-linear systems.
