+++ resonance/tests/test_systems.py
@@ -1,0 +1,39 @@
+import pytest
+
+from ..systems import _ParametersDict, _MeasurementsDict, _CoordinatesDict
+
+
+def test_nonvalid_parameters_key():
+    p = _ParametersDict({})
+    with pytest.raises(ValueError):
+        p['ben stiller'] = 12.0
+    with pytest.raises(ValueError):
+        p['time'] = 12.0
+    p['a'] = 1.0
+    if 'a' in p:
+        pass
+    del p['a']
+
+
+def test_setting_measurements_item():
+    m = _MeasurementsDict({})
+    # not allowed to set measurements
+    with pytest.raises(ValueError):
+        m['a'] = 12.0
+
+
+def test_setting_coordinates_item():
+    m = _CoordinatesDict({})
+    with pytest.raises(ValueError):
+        m['a '] = 12.0
+
+    with pytest.raises(ValueError):
+        m['time'] = 12.0
+
+    m['first_key'] = 12.0
+    with pytest.raises(ValueError):
+        m['second_key'] = 12.0
+
+    if 'first_key' in m:
+        pass
+    del m['first_key']
